requirements.txt
accelerate==0.27.2
beautifulsoup4==4.12.3
boto3==1.34.49
botocore==1.34.49
bpemb==0.3.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
conllu==4.5.3
contourpy==1.2.0
cycler==0.12.1
Deprecated==1.2.14
filelock==3.13.1
flair==0.13.1
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.1.3
gdown==5.1.0
gensim==4.3.2
huggingface-hub==0.20.3
idna==3.6
Janome==0.5.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pickle-mixin==1.0.2
pillow==10.2.0
pptree==3.1
protobuf==4.25.3
psutil==5.9.8
pycorenlp==0.3.0
pylzma==0.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch_revgrad==0.2.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
segtok==1.5.11
semver==3.0.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
sqlitedict==2.1.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformer-smaller-training-vocab==0.3.3
transformers==4.38.1
triton==2.2.0
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.2.13
Wikipedia-API==0.6.0
wrapt==1.16.0
